Changed behavior on unknown patterns to allow usage of custom patterns with the wrapper.
Before TryGetPattern was calling WrapObjectAsPattern, which in turn asked Schema for pattern info. Schema is internal and not extensible, hence UIA COM Wrapper effectively can't be used with patterns outside of set of patterns hardcoded in Schema. This changes the behavior: if we don't know how to wrap pattern - just return it as it is and let caller deal with consequences. Actually, it seems fine for custom patterns as author of custom pattern should know what he needs to do in such case.