Replaced single _identifierDirectory with one Dictionary<int, AutomationIdentifier> for each AutomationType.
Custom patterns' IDs intersect with ControlTypes IDs. I've encountered a situation where registration of my custom pattern returned id equal to ID of existing ControlType (Button) (this ID is 50000). Hence it is not correct in general to put all IDs to single Hashtable. This change makes the declaration into code.