Solving casing issues part 1: temporarily removed both folders where different casing is present
Casing issues are important to solve, because GitHub is case sensitive when you try to load raw source files. And GitLink indexes sources looking at current folder structure, hence will try to get all files with the same casing and some files won't be loaded.
For example, consider two files A/1.txt and a/2.txt. If current folder is lowercase - GitLink will put URL of 1.txt in lowercase as well. Otherwise it will put 2.txt URL in upper-case. Both are incorrect unfortunately.

The easiest solution it seems, is to remove everything in one commit and add everything back in the next one, but with correct casing.